슬랙 정보 관리 방식 변경 - 슬랙봇 설치 콜백 API 추가, 채널 정보 조회 API 반환 값에 봇 설치 링크 추가

+++ src/main/java/com/example/weaponMaster/modules/slack/constant/UserSlackNoticeType.java
@@ -2,5 +2,5 @@ package com.example.weaponMaster.modules.slack.constant;
 
 public class UserSlackNoticeType {
 
-    public static final int AUCTION_NOTICE  = 1; // 경매 판매 알림
+    public static final Byte AUCTION_NOTICE  = 1; // 경매 판매 알림
 }
